requirements.txt
numpy >= 1.19.0
opencv-python >= 4.5.0
matplotlib >= 3.3.0
pillow >= 8.0.0
tensorflow >= 2.4.0
#tensorflow-examples @ git+https://github.com/tensorflow/examples.git
